Some attempts at trying to move screen recording to Android. It doesn't work for a few reasons: 1. Screen recording doesn't seem to work from within the app process without root access 2. Granting root access requires an emulator with root shell privileges 3. Emulators with root shell privileges don't seem to have the correct codec setups to allow screen recording (even when used from adb directly)
I think this is possible to do, much more investigation work is needed
to get it to work. For now, we'll need to record the videos outside of
the process.

For future reference: consider using run-as with adb shell to gain
access to the internal directory storage where app probably needs to
store videos (though this needs to work against the automatic pm clear
happening with instrumentation tests, which is why Q-compatible external
storage might be better for non-rooted devices).