Create emulator & AVD manually.
This should provide much more flexibility both in setting up the
specific emulator that we need, and speed up the run slightly since we
can kick off starting the emulator and building the tests at the same
time.

This also disables all other workflows to speed up debugging.

Note that this was done since the way
ReactiveCircus/android-emulator-runner was evaluating its script was
highly incompatible with environment variables, making configurations
much harder.

Note that this won't actually trigger the emulator tests; those need to
be re-enabled in a subsequent commit. This is meant for
debugging/investigation purposes.