Initial commit for algeraic expression support.
This commit introduces:
1. Lexer for math expressions with single-letter variables.
2. Math expression parsing (no functions or unary support yet).
3. Polynomial computation engine (division not yet working, and no
collecting like terms or general form computation).

This commit does not include (but these may be needed in future commits):
1. Evaluation (though this is pretty easy to do later on).
2. Explicit equality/equivalence checks for expressions (equality may be
possible with an exact tree match, maybe with some allowances for
certain equivalence deviations. Equivalence can be done by comparing the
resulting polynomials for the two expressions also converted to general
form).
3. Partial expression tree matching (I'm not entirely sure how to solve
this yet. Simply looking for a subtree won't work since the tree's order
is based on order of operations which can change based on operations
surrounding a particular sub-expression).
4. Equation matching (I *think* this can be done by computing an
expression that = 0 by subtracting the left-hand side with the
right-hand side, then performing the polynomial-level checks. This will
also require additional parsing support for the right-associative
equivalence operator: =).

Extra noteworthy parts:
1. The lexer + parser are pretty efficient. Polynomial computation is
expensive due to its inherent complexity.
2. During polynomial computation, fractions/whole numbers are kept as
long as possible to avoid floats being passed around.

Note that this commit is introducing the system in a pretty rough state.
The final code organization, documentation, and even tests aren't yet
finalized here. This commit should only be considered a
proof-of-concept.