Fix & test math expression parsing.
This fixes a few issues with the core parsing, largely refactors it to
be more elegant, robust, and streamlined, adds documentation, adds
significant testing, and properly adds two new features: unary negation
and implied multiplication.

All of the different feature aspects of expression parsing, the
features above, and error cases are being thoroughly tested.

None of the expression extention library or polynomial functionality is
yet being tested.

Finally, 'NEGATION' was renamed to 'NEGATE' in MathUnaryOperation to
have parity with the operator names in MathBinaryOperation.