Add support for unpublished topics (#2303)
* Add support for unpublished topics.

Marking a topic as unpublished is the key way to indicate that it may be
coming in the future. The proto modeling for this is a oneof to be a bit
more future-proof, but currently there are only two states: can be
played, cannot be played. Conceivably, we may add future states to help
display potential release dates (e.g. quarters, specific time, etc.).

* Update topic.proto

Fix TopicPlayAvailability indexes.