Fix all tests.
This commit includes a lot, but in summary it ensures all tests that
rely on OppiaClock are now both passing and leveraging the clock
effectively (e.g. by properly utilizing each of the modes).

Some other structural changes to the clock itself to support super
specicfic time setting for certain tests.