Consolidate timing with a new fake oppia clock (#2644)
* Introduce new OppiaClock interface & fake.

Also, migrate (most) tests over to the new system. Fake tests still need
to be added, the branch needs to be brought up-to-date, and some other
minor fixes may be needed.

* Fix all tests.

This commit includes a lot, but in summary it ensures all tests that
rely on OppiaClock are now both passing and leveraging the clock
effectively (e.g. by properly utilizing each of the modes).

Some other structural changes to the clock itself to support super
specicfic time setting for certain tests.

* Redo tests for StoryProgressTestHelper.

* Simplify StoryProgressTestHelper's API & docs.

* Add tests for FakeOppiaClockTest.

Rename the fake clock's time mode enum constants.

* Lint fixes.