Fix affected target computation breaking CI workflows when no targets are affected (#2696)
* See if the extra newline is the source of issue.

* Temporarily disable unrelated checks.

* Conditionally run tests.

While Bazel can run an empty list of tests, GitHub actions always
expects at least 1 entry for matrix-dependent actions.

This is based on
https://github.community/t/empty-matrix-fails-workflow/128736/2. It's
not yet clear if this will properly allow check_test_results to pass if
the workflow is skipped. That will need to be tested.

* Fix Bash syntax error.

* Update test results job to pass without tests run.

* Update main.yml

Reenable jobs.

* Update compute_affected_tests.sh

Undo removing the newline since it's not actually needed.

* Re-remove newline since it is needed.

I forgot that we're performing an empty check, so the extra newline does
break it.