Reorganize GitHub Actions workflows (#2694)
* Move linter tests to the static checks workflow.

* Create build_tests.yml

Introduce a separate workflow file for build tests (which currently only includes the workflow for building the Bazel binary.

* Update unit_tests.yml

Clarify purpose of the unit tests workflow file in a documentation comment.

* Update main.yml

Repurpose main.yml to be specific to Gradle. We'll remove main.yml when we drop Gradle support (which will conveniently lead to having better segmented workflow files for GitHub Actions).