Fix LaTeX SVG rendering alignment.
Most of the work is being done in our monkeypatched androidsvg library,
but there are some minor changes needed in Oppia as well. This utilizes
properly computed SVG width/height, and factors in vertical alignment.
It also properly interops with Android's own spacing for inline
drawables so that spaces around the LaTeX image are represented
correctly.