Fix image loading in a couple places.
Particularly:
1. This fixes an issue with our SVG pipeline. Before, we were always
using our SVG rendering even for local drawables, but now we're actually
setting up the pipeline correctly to only load via androidsvg when we
intend to. This fixes profile avatar loading (which was broken due to
the new text svg pipeline that was added). This was an issue with the
existing glide module.

2. This finishes the proto version of promoted story creation so those
load correctly.

3. This reintroduces support for lesson thumbnails by creating separate
block/in-line image loading pipelines so that both can be loaded and
arranged correctly.

HTML-based images haven't yet been reintroduced.