Fix crash & images.
This fixes a crash introduced in an earlier commit due to now-unneeded
code that made an incorrect presumption of Glide's loading pipeline.
This also restores in-lesson image loading support & makes it work a bit
better than before.

This commit marks the end of known image issues; everything
image-related seems to be working correctly & well as needed for alpha.