Fix part of fixing broken interactions.
This updates interaction classifiers to support new types for text
input, drag and drop, and item selection/multiple choice. It partly
updates the UI portions (mainly for item selection), but more work is
needed both for item selection and drag and drop. Text input and
multiple choice should now be working.

Most of this work is due to structural changes in Oppia's backend in
preparation for supporting written translations for answers.