[RunAllTests] Fix #3018, #1675, #2931, #1600, and part of #2642: Enable data module in Bazel (#3025)
* Enable Moshi support in Bazel.

This includes some reorganizing of the data module, but it introduces
support for generating GAE models through Moshi. Corresponding mock
services & tests are still failing, but //data & //:oppia build.

* Make mock services build for testing module.

* Add working data tests to Bazel build.

This also fixes a regression from an earlier commit that led to the
Moshi objects not actually being generated correctly, so the data
service tests are good proof that Moshi is being used correctly.

* Remove TODOs for #1675.

* Lint fixes.

* Bazel lint fix.

Avoid relying on native java_library.

* Fix broken Gradle build.

This includes explicitly specifying okhttp version since Gradle was
previously randomly picking one.

* Add debug details for affected tests script.

* Lint fixes.

* Attempt to force Bazel cache to clear.

This is an attempt to work around a theorized OOM happening when trying
to compute the sibling build graph for changes build files.

* Post-merge fix.

* Fix another post-merge failure.

This regressed in a recent PR likely due to other Bazel failures hiding
it in CI.

* Upgrade to Robolectric 4.4.

This also moves off of Oppia's fork of robolectric-bazel.

* Switch TopicRevisionFragmentTest to test image loaders.

The test should have been using them before, but this might have been
missed. The test image loaders are necessary in the Robolectric test
environment, and something about Robolectric 4.4 + Bazel specifically
exposed this.