[RunAllTests] Fix #3510: Set up Android SDK hermetically in GitHub Actions to workaround #3024 (#3513)
* Refactor common build env setup to local action.

This attempts to move all common build environment setup bits to a
common local action using GitHub composite actions per
https://github.community/t/path-to-action-in-the-same-repository-as-workflow/16952/2,
https://github.blog/changelog/2020-08-07-github-actions-composite-run-steps/,
and
https://docs.github.com/en/actions/creating-actions/creating-a-composite-run-steps-action.

This does not include any actual changes to the CI environment yet other
than temporarily disabling the other expensive checks as part of testing
these changes.

* Update local action to only use shell commands.

* Set up JDK 9 manually for build tests.

* Fix Java version checking.

OpenJDK can use "9.0" as the version string instead of "1.9".

* Fix Java exit from version check.

* Fix incorrect check.

* Add Bazel setup.

* Fix syntax error & add Bazel version check.

* Add SDK setup bits.

* Avoid mv warning/error.

* Add debug lines.

* Attempt to fix pathing for new SDK.

* Attempt to workaround unexpected failure from sdkmanager.

* Add debug lines.

* Try to run license bit in a subshell to avoid exit

* Add codeowners, TODO, cleanup, and integrate.

This integrates the new action with Bazel tests in addition to the
build. While the tests work fine with the current build tools, we should
be consistent in how the environment is set up for CI workflows.

* Make local action name clearer.

The new name specifically clarifies why this action isn't needed for the
Bazel-only script runs (since they don't require the Android SDK bits).