Introduce string descriptions.
The language code 'qq' is used to represent descriptions for strings
that will then be exported to Translatewiki for translation. These
descriptions help provide context to translators that they may not be
able to easily get from the raw untranslated text.

This PR introduces an initial set of descriptions. The rest of the app's
strings will need to be backfilled.

Further, this PR also marks app_name as not translatable (something I
noticed in passing).