Add new tokenizer & parser.
This introduces a new tokenizer that's based on sequences rather than a
buffer. It's considerably shorter & simpler than the previous version.

This also introduces a parser specific to numeric expressions. Through
tests, this has been verified to have correct associativity & order of
operations, except for implicit multiplication (which still needs to be
fixed). Exponentiation was specifically validated for associativity,
too.

This introduces a new model spec for MatchesUpToTrivialManipulations
(though more work is probably needed here).

This PR also introduces a custom Truth subject for testing the new
tokenizer and a bunch of test cases. MANY more test cases are yet to be
added, but this was used to verify the tokenizer at least at a high
level.

This PR also introduces custom Truth subjects & a DSL for verifying the
constructed AST from numeric expressions. This is much easier to read &
understand as compared with previous expression tests. ~29 test cases
were added to verify various parts of the parser, but it's expected
dozens more test cases will be added to thoroughly test the parser.

A basic evaluation system was added (with integer being a new Real
value, and various efforts to try and retain perfect precision during
evaluation). This was added to various test cases for the parser to
verify more concretely that the order of operations is correct per the
ordering of the AST.

Finally, the implementation & tests prove that the grammar is properly
LL(1). Future iterations could optimize parsing to be table-based if
performance ends up being a problem (though I doubt parsing will be the
bottleneck here since evaluation seems far more expensive).