Proof-of-concept: parser DSL.
This commit introduces a proof-of-concept Kotlin DSL to define the
numeric expression grammar generically. While it can be simplified some,
it doesn't seem like the more complex syntax (which ultimately doesn't
look close enough to EBNF to be exactly comparable) and extra
infrastructural complexity seems beneficial over the simpler,
hand-maintained recursive descent parser. Maybe this could be useful if
more languages need to be supported in the future.

Note that the NumericExpressionParserTest is failing in this commit due
to an infinite loop, so the parser infrastructure is observably broken
(but functionally complete).