Add math expression/equation parsing support.
This includes full error detection, and specific test suites for each
parsing case. Much revisement is needed in the tests, and some
additional issues may yet need to be fixed in the parser and/or
error-detection logic.

This is copied from #2173 with revisement & reduction since it's part of
a multi-PR split.