Add new math expression test lessons.
This includes adding JSON support for the three new interactions. It
also fixes a bug in ExplorationProgressController wherein certain
exceptions would be swallowed by a subsequent initialization issue.

The three new explorations and each of their classifiers are
demonstrated in a single new exploration that's been added to the first
test story of the first test topic.