Add workflow to show build stats.
This workflow is set up to build multiple binaries for both develop &
the local branch, then analyze and generate reports that are uploaded as
artifacts and added as a comment to the PR body.

This doesn't make a guarantee about exact differences since GitHub uses
a detached HEAD for the commit that may not exactly match.

This also introduces a new workflow for stats (that will initially only
contain optional workflows since they may be very expensive/time
consuming).