Tests + docs + build changes.
This updates the script to no longer require an Android SDK path to be
passed in since it now directly pulls in AAPT2 and directly performs the
work of apkanalyzer using its internal library rather than calling
through to the CLI library (since it's not exposed via
android_sdk_repository).

This also updates the report output to compute full comparisons for all
APKs (rather than just the universal APK) and include it as part of the
full summary (so that only 2 files are ever written regardless of the
number of profiles being compared).

This adds placeholder tests & documentation for all new utilities.

This commit includes substantial changes to the build system:
- A new dependency on https://github.com/oppia/archive-patcher
- A split between production, test, and script Maven dependencies (where
the license tools have been updated to only operate on production
dependencies since those are the ones being shipped to users)
- Two versions of Guava are now being pulled in since the scripts
require a JRE version of Guava rather than the Android version used in
production & test classes
- rules_kotlin has been upgraded to beta 3 (from alpha 2), but the
Kotlin version itself is stuck on 1.4 (to ensure interoperability with
previous builds). While this change doesn't upgrade Kotlin, it does
allow us to leverage better tooling per changes in rules_kotlin,
including enforcing strict deps.