Reimplement numeric expression tests.
This is almost a full replacement. The new tests are more structured and
intentional to cover key high-level concepts. More tests may be added in
the future, but this is a sensible initial test offering.

This also updates MathExpressionSubject to support checking specifically
for implicit multiplication (and it's now required for such cases since
explicit is otherwise assumed).